Add full and foss product flavors
