+++ app/src/foss/java/com/zionhuang/music/utils/Utils.kt
@@ -1,0 +1,5 @@
+package com.zionhuang.music.utils
+
+import com.zionhuang.music.MainActivity
+
+fun MainActivity.setupRemoteConfig() {}
