+++ app/src/full/java/com/zionhuang/music/utils/Utils.kt
@@ -1,0 +1,31 @@
+package com.zionhuang.music.utils
+
+import com.google.firebase.ktx.Firebase
+import com.google.firebase.remoteconfig.ConfigUpdate
+import com.google.firebase.remoteconfig.ConfigUpdateListener
+import com.google.firebase.remoteconfig.FirebaseRemoteConfigException
+import com.google.firebase.remoteconfig.ktx.remoteConfig
+import com.google.firebase.remoteconfig.ktx.remoteConfigSettings
+import com.zionhuang.music.MainActivity
+import kotlin.time.Duration.Companion.hours
+
+fun MainActivity.setupRemoteConfig() {
+    val remoteConfig = Firebase.remoteConfig
+    remoteConfig.setConfigSettingsAsync(remoteConfigSettings {
+        minimumFetchIntervalInSeconds = 12.hours.inWholeSeconds
+    })
+    remoteConfig.fetchAndActivate()
+        .addOnCompleteListener(this) { task ->
+            if (task.isSuccessful) {
+                latestVersion = remoteConfig.getLong("latest_version")
+            }
+        }
+    remoteConfig.addOnConfigUpdateListener(object : ConfigUpdateListener {
+        override fun onError(error: FirebaseRemoteConfigException) {}
+        override fun onUpdate(configUpdate: ConfigUpdate) {
+            remoteConfig.activate().addOnCompleteListener {
+                latestVersion = remoteConfig.getLong("latest_version")
+            }
+        }
+    })
+}
